+++ libraries/tlv-account-resolution/Cargo.toml
@@ -12,13 +12,10 @@ test-sbf = []
 
 [dependencies]
 bytemuck = { version = "1.13.1", features = ["derive"] }
-num-derive = "0.3"
-num-traits = "0.2"
-num_enum = "0.6.1"
 solana-program = "1.16.1"
 spl-discriminator = { version = "0.1", path = "../discriminator" }
+spl-program-error = { version = "0.1.0", path = "../program-error" }
 spl-type-length-value = { version = "0.1", path = "../type-length-value" }
-thiserror = "1.0"
 
 [dev-dependencies]
 solana-program-test = "1.16.1"
